Modified the `Config` trait so it is now in charge of returning a ready-to-use `DoHProvider`.
`CLI` implements this by parsing the command line and, based on the given `--protocol` option, sees if a `--provider` is available (and logs a useful error if it doesn't).
This ensures that once we have got a usable `Config`, we can rely on it for the application to be fully functional.

I have also done a bit of work on the implementation of `Debug` for `DoHProvider`, so at runtime, if we log what provider was selected, the output is indicative enough.